Factor annotation and let id parsing through the same parser
This is a very trivial refactoring.

In an annotated let, the names in the annot and the let must match,
which we check in mkDecl; and the token parsers themselves are both
Fixpoint's lowerIdP, but in annot we reference FP.lowerIdP directly
and plainDecl's binder factors through Common/Parse.hs:identifier.
It makes no difference of course because identifier=FP.lowerIdP, but
if someone is playing with changing 'identifier' in Common/Parse.hs
then he will be confused by the difference.